A lot of things
So we did things like uhhh the umm some UI stuff and umm uhh text with who the player is, uhh the music as well. Um. UI is a little broken that's ok. yeah we changed that one highlight imag'es offset by half a unit. incredible. Started scene progression. Thanks for coming to my ted tlak.